+++ ts-client/src/dlmm/helpers/index.ts
@@ -1,4 +1,5 @@
 import { AnchorProvider, BN, EventParser, Program } from "@coral-xyz/anchor";
+import { IdlDiscriminator } from "@coral-xyz/anchor/dist/cjs/idl";
 import {
   ASSOCIATED_TOKEN_PROGRAM_ID,
   NATIVE_MINT,
@@ -11,7 +12,6 @@ import {
   getAssociatedTokenAddressSync,
   getMint,
 } from "@solana/spl-token";
-import { LBCLMM_PROGRAM_IDS, SCALE_OFFSET } from "../constants";
 import {
   Cluster,
   ComputeBudgetProgram,
@@ -20,6 +20,14 @@ import {
   SystemProgram,
   TransactionInstruction,
 } from "@solana/web3.js";
+import {
+  LBCLMM_PROGRAM_IDS,
+  MAX_BINS_PER_POSITION,
+  SCALE_OFFSET,
+  U64_MAX,
+} from "../constants";
+import IDL from "../dlmm.json";
+import { LbClmm } from "../idl";
 import {
   AccountName,
   Bin,
@@ -33,25 +41,23 @@ import {
   PresetParameter,
   PresetParameter2,
 } from "../types";
-import { Rounding, mulShr, shlDiv } from "./math";
 import {
-  getSimulationComputeUnits,
   MAX_CU_BUFFER,
   MIN_CU_BUFFER,
+  getSimulationComputeUnits,
 } from "./computeUnit";
-import IDL from "../dlmm.json";
-import { LbClmm } from "../idl";
-import { IdlDiscriminator } from "@coral-xyz/anchor/dist/cjs/idl";
+import { Rounding, mulShr, shlDiv } from "./math";
+import { LiquidityStrategyParameters } from "./rebalance";
 
-export * from "./derive";
 export * from "./binArray";
-export * from "./weight";
+export * from "./derive";
 export * from "./fee";
-export * from "./weightToAmounts";
-export * from "./strategy";
 export * from "./lbPair";
 export * from "./positions";
 export * from "./rebalance";
+export * from "./strategy";
+export * from "./weight";
+export * from "./weightToAmounts";
 
 export function chunks<T>(array: T[], size: number): T[][] {
   return Array.apply(0, new Array(Math.ceil(array.length / size))).map(
@@ -349,4 +355,62 @@ export function capSlippagePercentage(slippage: number) {
   }
 
   return slippage;
+}
+
+export function getAndCapMaxActiveBinSlippage(
+  slippagePercentage: number,
+  binStep: number,
+  maxActiveBinSlippage: number
+) {
+  return slippagePercentage
+    ? Math.ceil(slippagePercentage / (binStep / 100))
+    : maxActiveBinSlippage;
+}
+
+export function getBinCount(minBinId: number, maxBinId: number) {
+  return maxBinId - minBinId + 1;
+}
+
+export function getSlippageMaxAmount(amount: BN, slippage: number) {
+  return slippage == 100 ? U64_MAX : amount.muln(100 + slippage).divn(100);
+}
+
+export function getSlippageMinAmount(amount: BN, slippage: number) {
+  return amount.muln(100 - slippage).divn(100);
+}
+
+export function getPositionCountByBinCount(binCount: number) {
+  return Math.ceil(binCount / MAX_BINS_PER_POSITION.toNumber());
+}
+
+export function resetUninvolvedLiquidityParams(
+  minDeltaId: BN,
+  maxDeltaId: BN,
+  favorXInActiveId: boolean,
+  params: LiquidityStrategyParameters
+) {
+  const endBidSideDeltaId = favorXInActiveId ? new BN(-1) : new BN(0);
+  const startAskSideDeltaId = endBidSideDeltaId.addn(1);
+
+  let x0 = params.x0;
+  let y0 = params.y0;
+  let deltaX = params.deltaX;
+  let deltaY = params.deltaY;
+
+  if (maxDeltaId.lte(endBidSideDeltaId)) {
+    deltaX = new BN(0);
+    x0 = new BN(0);
+  }
+
+  if (minDeltaId.gte(startAskSideDeltaId)) {
+    deltaY = new BN(0);
+    y0 = new BN(0);
+  }
+
+  return {
+    x0,
+    y0,
+    deltaX,
+    deltaY,
+  };
 }